skills/freecad: factor out boilerplate across tests and scripts
Extract shared code to reduce ~215 lines of duplication:

- freecad_helpers.py: shared in-container helpers (log, pump,
  wait_for_view, init_gui) mounted into Docker alongside scripts.
  Added to skill_package for agent deployment.

- util/testing/otel_tracing.py: shared OTel test profiling
  (TracingConfig, span_to_dict) extracted from props/testing. Both
  skills/freecad and props now import from //util/testing.

- testing/compare.py: add assert_png_equal() for pixel-diff golden
  file comparison (was duplicated in test_render_3d + test_render_dxf).

- conftest.py: XVFB_CMD and FREECAD_HELPERS constants shared across
  test files.
