+++ skills/freecad/BUILD.bazel
@@ -10,6 +10,7 @@ skill_package(
         "build_compound.py",
         "build_cube_with_hole.py",
         "export_page.py",
+        "freecad_helpers.py",
         "parametric_sketch.py",
         "render_dxf.py",
         "render_fcstd.py",
@@ -48,9 +49,8 @@ py_library(
     deps = [
         "//util:oci",
         "//util/testing:container_logs",
-        "//util/testing:undeclared_outputs",
+        "//util/testing:otel_tracing",
         "@pypi//opentelemetry_api",
-        "@pypi//opentelemetry_sdk",
         "@pypi//pytest",
         "@pypi//pytest_bazel",
     ],
@@ -86,9 +86,9 @@ py_test(
     deps = [
         ":conftest",
         ":render_dxf",
+        "//skills/freecad/testing:compare",
         "//util/bazel:runfiles",
         "@pypi//opentelemetry_api",
-        "@pypi//pillow",
         "@pypi//pytest_bazel",
     ],
 )
@@ -99,6 +99,7 @@ py_test(
     srcs = ["test_render_3d.py"],
     data = [
         "build_cube_with_hole.py",
+        "freecad_helpers.py",
         "golden/cube_with_hole.png",
         "render_fcstd.py",
         ":freecad_test",
@@ -106,12 +107,12 @@ py_test(
     requires_docker = True,
     deps = [
         ":conftest",
+        "//skills/freecad/testing:compare",
         "//util:oci",
         "//util/bazel:runfiles",
         "//util/testing:container_logs",
         "//util/testing:undeclared_outputs",
         "@pypi//opentelemetry_api",
-        "@pypi//pillow",
         "@pypi//pytest",
         "@pypi//pytest_bazel",
     ],
@@ -123,6 +124,7 @@ py_test(
     srcs = ["test_parametric_sketch.py"],
     data = [
         "export_page.py",
+        "freecad_helpers.py",
         "golden/bracket.dxf",
         "golden/bracket.pdf",
         "golden/bracket.svg",
@@ -150,6 +152,7 @@ py_test(
     data = [
         "build_compound.py",
         "export_page.py",
+        "freecad_helpers.py",
         "golden/compound.dxf",
         "golden/compound.pdf",
         "golden/compound.svg",
