+++ skills/freecad/test_compound.py
@@ -7,7 +7,7 @@ import pytest
 import pytest_bazel
 from opentelemetry import trace
 
-from skills.freecad.conftest import FREECAD_TEST, freecad_exec
+from skills.freecad.conftest import FREECAD_HELPERS, FREECAD_TEST, XVFB_CMD, freecad_exec
 from skills.freecad.testing.compare import assert_dxf_equal, assert_pdf_equal, assert_svg_equal
 from util.bazel.runfiles import get_required_path
 from util.oci import load_oci_image
@@ -19,8 +19,6 @@ _EXPORT_SCRIPT = "_main/skills/freecad/export_page.py"
 _GOLDEN_DXF = "_main/skills/freecad/golden/compound.dxf"
 _GOLDEN_SVG = "_main/skills/freecad/golden/compound.svg"
 _GOLDEN_PDF = "_main/skills/freecad/golden/compound.pdf"
-
-_XVFB = 'xvfb-run -a -s \\"-screen 0 1024x768x24\\"'
 
 tracer = trace.get_tracer(__name__)
 
@@ -39,6 +37,7 @@ def compound_outputs(tmp_path_factory: pytest.TempPathFactory) -> Path:
             test_name="freecad-compound",
             command="sleep infinity",
             volumes=[
+                (str(get_required_path(FREECAD_HELPERS)), "/work/freecad_helpers.py", "ro"),
                 (str(get_required_path(_BUILD_SCRIPT)), "/work/build_compound.py", "ro"),
                 (str(get_required_path(_EXPORT_SCRIPT)), "/work/export_page.py", "ro"),
                 (str(tmp_path), "/output", "rw"),
@@ -46,9 +45,10 @@ def compound_outputs(tmp_path_factory: pytest.TempPathFactory) -> Path:
             docker_client_kw={"timeout": 120},
         ) as container,
     ):
-        freecad_exec(container, f'bash -c "OUTDIR=/output {_XVFB} freecadcmd /work/build_compound.py"')
+        freecad_exec(container, f'bash -c "OUTDIR=/output {XVFB_CMD} freecadcmd /work/build_compound.py"')
         freecad_exec(
-            container, f'bash -c "INPUT=/output/compound.FCStd OUTDIR=/output {_XVFB} freecadcmd /work/export_page.py"'
+            container,
+            f'bash -c "INPUT=/output/compound.FCStd OUTDIR=/output {XVFB_CMD} freecadcmd /work/export_page.py"',
         )
 
     out_dir = undeclared_outputs_dir() / "compound"
