+++ skills/freecad/test_render_dxf.py
@@ -8,16 +8,13 @@ from pathlib import Path
 
 import pytest_bazel
 from opentelemetry import trace
-from PIL import Image
 
 from skills.freecad.render_dxf import render_dxf
+from skills.freecad.testing.compare import assert_png_equal
 from util.bazel.runfiles import get_required_path
 
 _GOLDEN_DXF = "_main/skills/freecad/golden/compound.dxf"
 _GOLDEN_PNG = "_main/skills/freecad/golden/compound.png"
-
-# Maximum fraction of differing pixel channels tolerated.
-_MAX_DIFF_FRACTION = 0.02
 
 tracer = trace.get_tracer(__name__)
 
@@ -31,17 +28,7 @@ def test_render_compound(tmp_path: Path) -> None:
         render_dxf(golden_dxf, actual_png)
 
     with tracer.start_as_current_span("compare_golden_png"):
-        actual = Image.open(actual_png).convert("RGB")
-        golden = Image.open(golden_png_path).convert("RGB")
-        assert actual.size == golden.size, f"Size mismatch: {actual.size} vs {golden.size}"
-
-        a_data = actual.tobytes()
-        g_data = golden.tobytes()
-        differing = sum(1 for a, g in zip(a_data, g_data, strict=True) if a != g)
-        diff_fraction = differing / len(a_data)
-        assert diff_fraction <= _MAX_DIFF_FRACTION, (
-            f"Rendered PNG differs from golden by {diff_fraction:.1%} (threshold {_MAX_DIFF_FRACTION:.1%})"
-        )
+        assert_png_equal(actual_png, golden_png_path)
 
 
 if __name__ == "__main__":
