+++ util/testing/BUILD.bazel
@@ -12,6 +12,19 @@ py_library(
 )
 
 py_library(
+    name = "otel_tracing",
+    testonly = True,
+    srcs = ["otel_tracing.py"],
+    imports = ["../.."],
+    visibility = ["//visibility:public"],
+    deps = [
+        ":undeclared_outputs",
+        "@pypi//opentelemetry_api",
+        "@pypi//opentelemetry_sdk",
+    ],
+)
+
+py_library(
     name = "container_logs",
     testonly = True,
     srcs = ["container_logs.py"],
